requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.8.3
pydantic==2.5.0
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
aiofiles==23.2.1
httpx==0.27.0